Changed doubles to longs in report. Intermediate point: need to move to BigDecimal to get rid of magic 100's and 10000's all over the place

+++ branches/repr-expr/invextension/src/com/moneydance/modules/features/invextension/DateUtils.java
@@ -534,10 +534,10 @@ public final class DateUtils {
      */
 
 
-    public static double getExcelDateValue(int dateInt) {
+    public static long getExcelDateValue(int dateInt) {
         int dateStartInt = convertToDateInt(dateStart.getTime());
-        Integer daysBetwInt = getDaysBetween(dateStartInt, dateInt);
-        return daysBetwInt.doubleValue();
+        int daysBetwInt = getDaysBetween(dateStartInt, dateInt);
+        return daysBetwInt;
     }
 
     /**
